requirements.txt
Pillow
geopy
numpy
